iwlwifi: Use driver rssi stats to determine link RSSI
The beacon stats appear to be buggy for the secondary link
and/or maybe just because I'm disabling beacon filtering.
So calculate these in the driver instead, similar to how I
did per-vif averages in the past.

+++ drivers/net/wireless/intel/iwlwifi/mvm/coex.c
@@ -333,7 +333,7 @@ void iwl_mvm_bt_coex_update_link_esr(struct iwl_mvm *mvm,
 		return;
 
 	if (!iwl_mvm_bt_coex_calculate_esr_mode(mvm, vif,
-						(s8)link->beacon_stats.avg_signal,
+						link->beacon_stats.avg_signal,
 						link_id == iwl_mvm_get_primary_link(vif)))
 		/* In case we decided to exit eSR - stay with the primary */
 		iwl_mvm_exit_esr(mvm, vif, IWL_MVM_ESR_EXIT_COEX,
